requirements.txt
asgiref==3.7.2
Django==4.2.3
django-extensions==3.2.3
django-taggit==4.0.0
lorem==0.1.1
Markdown==3.4.3
psycopg2-binary==2.9.6
sqlparse==0.4.4
